Project information
This project basically represents a sub-department of a corporate holding. At this point, a home page and admin panel were used, and in this panel, the information belonging to the admin was displayed and sensitive information drawn from the database was displayed. In this project, dotnet 6.0 version is used and Entity Framework ORM, ASP.NET Core Web API, MVC, Consume, Filters, Aspect Oriented Programming, View Components etc. approaches have been used. A flexible library has been defined in the project that can be used independently in each project. In general, a service has been developed and it is a project that aims to consume it with MVC and has achieved this.